Integrate with `turbo:load`
Administrate `1.0.0` is poised to integrate with Turbo and Stimulus.
That means that plugins that utilize client-side JavaScript cannot
depend on `DOMContentLoaded` events in the same way as pre-1.0.0.

First, check for `window.Turbo`. If that exists, install the listener
for [`turbo:load`][] events.

If not, fall back to using `jQuery` to attach [`DOMContentLoaded`][]
events.

[`turbo:load`]: https://turbo.hotwired.dev/reference/events#turbo%3Aload
[`DOMContentLoaded`]: https://developer.mozilla.org/en-US/docs/Web/API/Document/DOMContentLoaded_event